map unsent messages (#60)
* map unsent messages

Map messages that contain "rp" and "otr" data in the summary information
to indicate that they've been unsent. I've quickly verified that unsent
messages contain both of these properties by (un)sending messages to
myself, and looking at past conversations with others. They only seem to
be present on unsent messages.

Arguably, you could also check for unsent messages by seeing if the
content of a message is null, but that seems less robust to me. I also
want to account for the possibility that a message can have a null
content while not being unsent, as I'm not terribly familiar with how
iMessage's data modeling invariants.

* Correct mistakenly truncated comment

* properly handle partially unsent messages

After discovering the structure to `otp` and `rp`, we can use it as a
breadcrumb trail of sorts to figure out where to insert "(user) unsent a
message" indicators.

* remove extra }