handle more partially unsent messages correctly (#61)
* handle more partially unsent messages correctly

Deal with unsent portions of messages in a more resilient way, which
enables correct handling of multiple adjacent unsent parts in the
middle, end, and beginning of messages. Instead of performing a naive
"otr.lo < frag.to" comparison (which is somewhat conceptually scuffed),
track the part attributes that iMessage provides when stepping over the
parsed message fragments. We can then detect gaps in these part numbers,
which lets us insert unsent message parts before proceeding with the
message fragment.

This requires some index rejiggering, but it's far more correct. This
improved handling also provides the correct indexes to Texts.

* use the right indexes for partial message parts

I got confused and assumed that the `index` property corresponds to the
index of the first character of that part within the entire string,
instead of simply the part's index within the parts so far.

* rename variable to avoid confusion

Match the `index` variable with the property it's actually used with to
avoid any potential confusion.