RustServer: sync messages before threads
this is necessary so the unread count doesn't start at 2, and mirrors
how e.g. cloud discord messages are synced. if you sync the thread
first, then the unread count is bumped unnecessarily when the message
comes in. by syncing the message first, the unread count in the renderer
gets bumped _immediately_ and the final unread count gets synced
afterwards, potentially fixing it. this seems to be what's intended.