[DESK-8051] state sync `unreadCount` and `lastReadMessageSortKey`, just like beeper (#26)
* thread: fetch unread counts from imsg db, map as `unreadCount`

* thread: don't map `isUnread` anymore

beeper desktop ThreadStore's `isUnread` is now computed (from
isMarkedUnread and unreadCount) and shouldn't be state synced anymore

* thread: map `c.last_read_message_timestamp` to `lastReadMessageSortKey`

this timestamp doesn't seem to always correspond to an existing
message's timestamp, but that seems OK for beeper desktop - the unread
line is still correctly placed.

* RustServer: poll for unread counts and last read message ts

* fix token pass

* RustServer: send state syncs for threads turning read

* RustServer: track unread state for all chats, not just unread

This is needed for proper handling of chats that receive new messages,
but never actually turn unread (can happen when you have the chat
focused on another device).

* RustServer: sync messages before threads

this is necessary so the unread count doesn't start at 2, and mirrors
how e.g. cloud discord messages are synced. if you sync the thread
first, then the unread count is bumped unnecessarily when the message
comes in. by syncing the message first, the unread count in the renderer
gets bumped _immediately_ and the final unread count gets synced
afterwards, potentially fixing it. this seems to be what's intended.