mappers: fall back to entire message instead of text part when one isn't found (#31)
I haven't looked into the root cause of this, but this at least prevents
bogus Message objects (e.g. with undefined `id` or `timestamp`) to hit
the renderer. In my case, this resulted in a thread that would re-appear
to the top of the inbox every time you loaded the app, and it could
never be archived.

Fixes DESK-8329.