[DESK-9302] imessage: have SwiftServer absorb RustServer (#49)
* update `swift-tools-version` to 5.8

* sqlite: add basic wrapper library

* imdb: init

* sqlite: flesh out wrapper

* imdb: clean up

* sqlite: accept flags when opening, no longer RO

* package: add soft symbol resolution linker flags

so we can build with SPM (only in dev, for convenience)

* sqlite: add tests

* imdb: add unread queries

* sws: add hashing, add tests for them

* imdb/tbench: add demo for unread states

* imdb: add cached statement helper

* imdb: comments

* sqlite/statement: add `mapRowsUntilDone`

* imdb: add support for reading chat updates

* imdb: break out date helpers

* deps: use `node-swift` with 15.4 TLS hack

this is only needed if you are consuming the `SwiftServer.node` module
directly (i.e. not from Beeper Desktop), since the fixed node-swift
would automatically take priority there. making the binding directly
usable helps with testing

* hashing: expose `Hasher`s to Node

* api: don't use hashers directly, use from `hashing.ts`

* hashing: lint

* poller: poll unread states

* imdb: add `ChatRef` convenience initializer

* sqlite: make `.as(String.self)`'s return optional

oops, the return value was an implicitly unwrapped optional

* imdb/updates: use `ChatRef`, don't always assume one is read

when our chats get updated, there won't necessarily be a message with a
non-zero `date_read` in there. make sure we don't send a weird value
when this happens

easily replicable with two iphones texting each other in a normal convo,
where one of them is connected to beeper

* poller: refresh threads on chat updates

this is a bit complicated by the fact that we need the `onEvent`
callback _and_ initial cursor seed values in order to begin polling.
continue doing the hack to wait until we have both

* sqlite/statement: separate stepping methods

* move hex string utilities to foundation

* poller: split up main portions

* ` `

* api: s/RustServer/SwiftServer

* hasher: throw when not found to match rust behavior

* poller: fix `ChatRef.description` using the wrong hasher

* consume hashers from node side

* goodbye crab

* 0.9.0