eclipsing: flip electron frame before eclipsing
with help from the debugger:

`NSWindow` frame coordinate space has the origin at the bottom left but
`NSScreen` has it at the top left. we weren't correcting for this, so
the further away you'd have Beeper from the "equator" of your display,
the worse the eclipsing would end up behaving.