use FSEvents to resurrect dead `chat.db-wal` watchers
now that we can watch `~/Library/Messages` with FSEvents, leverage this
to respond to the WAL/db file being recreated, which happens when SQLite
checkpoints it into `chat.db`.

(in actuality, don't watch for WAL/db file events specifically - let's
just watch for events occurring directly under `~/Library/Messages` and
resurrect our file watchers when we notice something. this is to reduce
work on fseventsd's side and to keep things lean.)

to determine whether we actually have to revive a WAL/db file watcher,
check the number of hard links on the fd via `fstat` -- it'll zero out
once SQLite finishes checkpointing, which we can reliably leverage in
order to respond to it. (this also means that unrelated file events
directly under `~/Library/Messages` are just ignored, pretty much.)

to simplify things, watchers no longer expose their own `Topic`, instead
you just set a callback on them that is invoked as you'd expect. this
allows clients to multiplex observed events on their own, which is
vastly simpler compared to having to `merge` multiple disparate `Topic`
instances via AsyncAlgorithms or something.

this should be the final piece of the puzzle to fix BRI-33020