remove `openAfter` usage
this might not be strictly necessary but it may help with avoiding the
Messages app from jumping around a ton of chats. it may also speed
things up as a side effect.

since the idle callback is called continuously over time, the controller
code no longer has to know about preserving the activity observer
anymore